+++ src/BeatPulse.AzureServiceBus/AzureServiceBusQueueLiveness.cs
@@ -47,7 +47,7 @@ namespace BeatPulse.AzureServiceBus
             {
                 _logger?.LogWarning($"The {nameof(AzureServiceBusQueueLiveness)} check fail for {_connectionString} with the exception {ex.ToString()}.");
 
-                return LivenessResult.UnHealthy(ex, showDetailedErrors: context.ShowDetailedErrors);
+                return LivenessResult.UnHealthy(ex);
             }
         }
     }