+++ src/BeatPulse.Network/FtpLiveness.cs
@@ -49,7 +49,7 @@ namespace BeatPulse
             {
                 _logger?.LogWarning($"The {nameof(FtpLiveness)} check fail with the exception {ex.ToString()}.");
 
-                return LivenessResult.UnHealthy(ex, showDetailedErrors: context.ShowDetailedErrors);
+                return LivenessResult.UnHealthy(ex);
             }
         }