+++ src/BeatPulse.Network/SmtpLiveness.cs
@@ -56,7 +56,7 @@ namespace BeatPulse.Network
             {
                 _logger?.LogWarning($"The {nameof(SmtpLiveness)} check fail with the exception {ex.ToString()}.");
 
-                return LivenessResult.UnHealthy(ex, showDetailedErrors: context.ShowDetailedErrors);
+                return LivenessResult.UnHealthy(ex);
             }
         }
     }