+++ src/BeatPulse.Sqlite/SqliteLiveness.cs
@@ -45,7 +45,7 @@ namespace BeatPulse.Sqlite
             {
                 _logger?.LogWarning($"The {nameof(SqliteLiveness)} check fail for {_connectionString} with the exception {ex.ToString()}.");
 
-                return LivenessResult.UnHealthy(ex, showDetailedErrors: context.ShowDetailedErrors);
+                return LivenessResult.UnHealthy(ex);
             }
         }
     }