+++ src/BeatPulse.UI/Configuration/BeatPulseSettings.cs
@@ -25,5 +25,6 @@ namespace BeatPulse.UI.Configuration
         public string Name { get; set; }
         public string Uri { get; set; }
         public string Payload { get; set; }
+        public string RestoredPayload { get; set; }
     }
 }