+++ src/BeatPulse.UI/Core/LivenessFailureNotifier.cs
@@ -24,43 +24,65 @@ namespace BeatPulse.UI.Core
         {
             foreach (var webHook in _settings.Webhooks)
             {
-                if (webHook.Uri == null || !Uri.TryCreate(webHook.Uri, UriKind.Absolute, out Uri webHookUri))
+                var targetWebHook = new WebHookNotification()
                 {
-                    _logger.LogWarning($"The web hook notification uri is not stablished or is not an absolute Uri ({webHook.Name}). Set the webhook uri value on BeatPulse setttings.");
+                    Name = webHook.Name,
+                    Uri = webHook.Uri                    
+                };
 
-                    continue;
-                }
+                targetWebHook.Payload = webHook.Payload
+                    .Replace(BeatPulseUIKeys.LIVENESS_BOOKMARK, livenessName)
+                    .Replace(BeatPulseUIKeys.FAILURE_BOOKMARK, content);
 
+                await NotifyWebhookPayLoad(targetWebHook, targetWebHook.Payload);
+            }
+        }
+
+        public async Task NotifyLivenessRestored(string livenessName, string context)
+        {
+            foreach (var webHook in _settings.Webhooks)
+            {
+                var targetWebHook = new WebHookNotification()
+                {
+                    Name = webHook.Name,
+                    Uri = webHook.Uri                 
+                };
+
+                targetWebHook.RestoredPayload = webHook.RestoredPayload
+                    .Replace(BeatPulseUIKeys.LIVENESS_BOOKMARK, livenessName);
+
+                await NotifyWebhookPayLoad(targetWebHook, targetWebHook.RestoredPayload);
+            }
+        }
+
+        private async Task NotifyWebhookPayLoad(WebHookNotification webHook, string payloadContent)
+        {
+            if (webHook.Uri == null || !Uri.TryCreate(webHook.Uri, UriKind.Absolute, out Uri webHookUri))
+            {
+                _logger.LogWarning($"The web hook notification uri is not stablished or is not an absolute Uri ({webHook.Name}). Set the webhook uri value on BeatPulse setttings.");
+
+                return;
+            }
+
+            try
+            {
                 using (var httpClient = new HttpClient())
                 {
-                    var targetWebHook = new WebHookNotification()
+                    var payload = new StringContent(payloadContent, Encoding.UTF8, BeatPulseUIKeys.DEFAULT_RESPONSE_CONTENT_TYPE);
+                    var response = await httpClient.PostAsync(webHookUri, payload);
+
+                    if (!response.IsSuccessStatusCode)
                     {
-                        Name = webHook.Name,
-                        Uri = webHook.Uri,
-                        Payload = webHook.Payload
-                    };
-
-                    targetWebHook.Payload = targetWebHook.Payload
-                        .Replace(BeatPulseUIKeys.LIVENESS_BOOKMARK, livenessName)
-                        .Replace(BeatPulseUIKeys.FAILURE_BOOKMARK, content);
-
-                    var payload = new StringContent(targetWebHook.Payload, Encoding.UTF8, BeatPulseUIKeys.DEFAULT_RESPONSE_CONTENT_TYPE);
-
-                    try
-                    {
-                        var response = await httpClient.PostAsync(webHookUri, payload);
-
-                        if (!response.IsSuccessStatusCode)
-                        {
-                            _logger.LogError($"The failure notification is not executed successfully for {webHook.Name} webhook. The error code is {response.StatusCode}.");
-                        }
-                    }
-                    catch (Exception exception)
-                    {
-                        _logger.LogError($"The failure notification for {webHook.Name} is not executed successfully.", exception);
+                        _logger.LogError($"The webhook notification has not executed successfully for {webHook.Name} webhook. The error code is {response.StatusCode}.");
                     }
                 }
+
             }
+            catch (Exception exception)
+            {
+                _logger.LogError($"The failure notification for {webHook.Name} has not executed successfully.", exception);
+            }
+
         }
     }
 }