+++ src/BeatPulse/BeatPulseKeys.cs
@@ -1,5 +1,9 @@
 ﻿namespace BeatPulse
 {
+
+    /// <summary>
+    /// BeatPulse constants.
+    /// </summary>
     public class BeatPulseKeys
     {
         public const string BEATPULSE_HEALTHCHECK_DEFAULT_ERROR_MESSAGE = "The beat pulse check is not working properly.";
@@ -12,5 +16,6 @@
         internal const string BEATPULSE_INVALIDPATH_REASON = "Invalid BeatPulse path.";
         internal const string BEATPULSE_TIMEOUT_MESSAGE = "Timeout";
         internal const string BEATPULSE_SERVICEUNAVAILABLE_REASON = "One or more liveness return a service unavailable result.";
+        internal const string BEATPULSE_OPTIONS_SETTING_KEY = "BeatPulseOptions";
     }
 }